+++ packages/common/src/visualizations/TableDataModel.ts
@@ -50,11 +50,11 @@ export class TableDataModel
             (acc, key) => ({
                 ...acc,
                 [key]: {
-                    visible: true, // FIXME: should this be true all the time?
+                    visible: this.config?.columns[key].visible ?? true,
                     reference: key,
-                    label: key,
-                    frozen: true,
-                    order: undefined,
+                    label: this.config?.columns[key].label ?? key,
+                    frozen: this.config?.columns[key].frozen ?? false,
+                    order: this.config?.columns[key].order,
                 },
             }),
             {},