+++ packages/frontend/src/components/DashboardTiles/DashboardSqlChartTile.tsx
@@ -2,13 +2,17 @@ import {
     ChartKind,
     isVizTableConfig,
     type DashboardSqlChartTile as DashboardSqlChartTileType,
+    type SqlChart,
 } from '@lightdash/common';
-import { IconAlertCircle } from '@tabler/icons-react';
+import { Box, Menu } from '@mantine/core';
+import { IconAlertCircle, IconFilePencil } from '@tabler/icons-react';
 import { useMemo, type FC } from 'react';
-import { useParams } from 'react-router-dom';
+import { useHistory, useParams } from 'react-router-dom';
 import { useSqlChartAndResults } from '../../features/sqlRunner/hooks/useSqlChartAndResults';
 import { SqlRunnerResultsRunner } from '../../features/sqlRunner/runners/SqlRunnerResultsRunner';
+import MantineIcon from '../common/MantineIcon';
 import SuboptimalState from '../common/SuboptimalState/SuboptimalState';
+import { type ResultsAndColumns } from '../DataViz/Results';
 import ChartView from '../DataViz/visualizations/ChartView';
 import { Table } from '../DataViz/visualizations/Table';
 import TileBase from './TileBase';
@@ -27,7 +31,33 @@ interface Props
  * Handle minimal mode
  * handle tabs
  */
-
+const DashboardOptions = ({
+    isEditMode,
+    data,
+}: {
+    isEditMode: boolean;
+    data: {
+        resultsAndColumns: ResultsAndColumns;
+        chart: SqlChart;
+    };
+}) => {
+    const history = useHistory();
+    return (
+        <Box>
+            <Menu.Item
+                icon={<MantineIcon icon={IconFilePencil} />}
+                disabled={isEditMode}
+                onClick={() =>
+                    history.push(
+                        `/projects/${data.chart.project.projectUuid}/sql-runner/${data.chart.slug}/edit`,
+                    )
+                }
+            >
+                Edit SQL chart
+            </Menu.Item>
+        </Box>
+    );
+};
 export const DashboardSqlChartTile: FC<Props> = ({
     tile,
     isEditMode,
@@ -97,6 +127,9 @@ export const DashboardSqlChartTile: FC<Props> = ({
             tile={tile}
             title={tile.properties.title || tile.properties.chartName || ''}
             {...rest}
+            extraMenuItems={
+                <DashboardOptions isEditMode={isEditMode} data={data} />
+            }
         >
             {data.chart.config.type === ChartKind.TABLE &&
                 isVizTableConfig(data.chart.config) && (