+++ packages/frontend/src/components/DataViz/transformers/useChart.ts
@@ -4,6 +4,7 @@ import {
     isVizCartesianChartConfig,
     isVizPieChartConfig,
     PieChartDataModel,
+    type PivotChartData,
     type VizCartesianChartConfig,
     type VizPieChartConfig,
 } from '@lightdash/common';
@@ -18,6 +19,7 @@ export const useChart = <T extends ResultsRunner>({
     projectUuid,
     limit,
     orgColors,
+    onPivot,
 }: {
     config?: VizCartesianChartConfig | VizPieChartConfig;
     resultsRunner: T;
@@ -25,6 +27,7 @@ export const useChart = <T extends ResultsRunner>({
     projectUuid?: string;
     limit?: number;
     orgColors?: string[];
+    onPivot?: (pivotData: PivotChartData) => void;
 }) => {
     const chartTransformer = useMemo(() => {
         if (config?.type === ChartKind.PIE) {
@@ -54,7 +57,13 @@ export const useChart = <T extends ResultsRunner>({
         [chartTransformer, config?.fieldConfig, projectUuid, limit],
     );
 
-    const transformedData = useAsync(getTransformedData, [getTransformedData]);
+    const transformedData = useAsync(async () => {
+        const data = await getTransformedData();
+        if (onPivot && data) {
+            onPivot(data);
+        }
+        return data;
+    }, [getTransformedData]);
 
     const chartSpec = useMemo(() => {
         if (!transformedData.value) return undefined;